Fix critical bugs and make some little changes (#13)
* Fix critical bugs and some make some little changes
* Fix bug when member can see file preview with disable `Show preview` toggle
* Fix upload modal rounded class
* Optimize code and super fast file delete
* Bump MyStorage version

+++ apps/backend/src/routes/delete.ts
@@ -17,13 +17,17 @@ export async function deleteFile(req: FastifyRequest & { body: { id: string } },
       prisma.member.findUnique({ where: { id: req.user.member_id } }),
     ]);
 
+    if (!upload) return await res.status(404).send({ error: true, text: "Данного файла больше не существует!" });
+
     const deleteParams = {
       Bucket: member.bucket,
       Key: upload.key,
     } as DeleteObjectRequest;
 
+    const uploadS3 = await getS3(req.user.member_id);
+
     const [deletedObjectS3, file] = await Promise.all([
-      (await getS3(req.user.member_id)).deleteObject(deleteParams).promise(),
+      uploadS3.deleteObject(deleteParams).promise(),
       prisma.upload.delete(findParams),
       redisPub.del(`${member.id}:${upload.key}`),
     ]);
@@ -38,20 +42,25 @@ export async function deleteFile(req: FastifyRequest & { body: { id: string } },
 
 export async function forceDeleteFile(req: FastifyRequest, res: FastifyReply) {
   try {
+    const uploadS3 = await getS3(req.user.member_id);
+
     const uploads = await prisma.upload.findMany({
       where: { memberID: req.user.member_id },
       select: { Member: { select: { bucket: true, id: true } }, key: true, id: true },
     });
 
     for (const upload of uploads) {
-      await Promise.all([
-        (await getS3(req.user.member_id)).deleteObject({ Bucket: upload.Member.bucket, Key: upload.key }).promise(),
-        prisma.upload.delete({ where: { id: upload.id } }),
-        redisPub.del(`${upload.Member.id}:${upload.key}`),
-      ]);
+      uploadS3.deleteObject({ Bucket: upload.Member.bucket, Key: upload.key }).promise();
+      redisPub.del(`${upload.Member.id}:${upload.key}`);
     }
 
-    return await res.status(200).send({ error: false, count: uploads.length });
+    const count = await prisma.upload.deleteMany({
+      where: {
+        id: { in: uploads.map((u) => u.id) },
+      },
+    });
+
+    return await res.status(200).send({ error: false, count: count.count });
   } catch (err) {
     console.error("forceDeleteFile error:", err);
